+++ local/docs/figures/utils/style.py
@@ -1,5 +1,7 @@
 import matplotlib.pyplot as plt
 
+from max_div._core.plotting.helpers import figure_style_path
+
 
 def set_docs_style():
-    plt.style.use("./local/docs/figures/docs.mplstyle")
+    plt.style.use(figure_style_path())